Reorganize src
