+++ src/html/panel.html
