+++ src/lib/PluginStorage.js
