+++ src/lib/ZipkinPanel.js
