+++ src/lib/ZipkinPlugin.js
