+++ src/lib/addNetworkEvents.js
