+++ src/lib/attachBeforeSendHeadersListener.js
