+++ src/lib/checkZipkinUI.js
